refactor(AutoNumbersMaker : List<Integer> numbers 전역변수 처리 시 여러 곳에서 변경 접근 가능하여 인스턴스변수 처리
